Some reorganization
