+++ app/ui-react/syndesis/src/modules/connections/components/ConnectionsWithToolbar.tsx
