+++ app/ui-react/syndesis/src/modules/connections/components/index.ts
@@ -1,1 +1,4 @@
 export * from './ConnectionCreatorBreadcrumbs';
+export * from './Connections';
+export * from './ConnectionsWithToolbar';
+export * from './WithConnectorForm';
