--- app/ui-react/syndesis/src/modules/connections/pages/ConnectionDetailsPage.css
@@ -1,11 +1,0 @@
-.connection-details-page__propertiesForm {
-  margin-left: auto; 
-  margin-right: auto;
-  width: 90%;
-}
-
-.connection-details-page__sectionTitle {
-  font-size: 1.25em;
-  font-weight: 400;
-  margin: 20px;
-}
