--- app/ui-react/syndesis/src/modules/connections/shared/index.ts
@@ -1,3 +1,0 @@
-export * from './Connections';
-export * from './ConnectionsWithToolbar';
-export * from './WithConnectorCreationForm';
